Move `retrieveRepository` to `Repository` companion object
